+++ rmk/src/keyboard_macro.rs
@@ -1,7 +1,7 @@
 use crate::keycode::KeyCode;
 
 // Default macro space size
-pub(crate) const MACRO_SPACE_SIZE: usize = 512;
+pub(crate) const MACRO_SPACE_SIZE: usize = 256;
 
 // Default number of keyboard macros
 pub(crate) const NUM_MACRO: usize = 8;